Solution to Input and Print an Integer

+++ Input and Print an Integer@@6788d5012db25efec946f939/code.c
@@ -1,16 +1,15 @@
 
-
 #include <stdio.h>
 
 int main() {
-    float num;  // Variable to store the floating-point number
-
+    int num;
+    
     // Taking input
-    scanf("%f", &num);
-
-    // Printing output with 2 decimal places
-    printf("You entered: %.2f\n", num);
-
+    scanf("%d", &num);
+    
+    // Printing output
+    printf("You entered: %d\n", num);
+    
     return 0;
 }
